rewrite luconverter
